RDS/layer0_ingress/web/server/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
bidict==0.21.2; python_version >= '3.6'
cachelib==0.3.0; python_version >= '3.6'
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4; python_version >= '3'
click==8.0.1; python_version >= '3.6'
contextlib2==21.6.0; python_version >= '3.6'
cryptography==3.4.8
dnspython==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
eventlet==0.30.2
flask-cors==3.0.10
flask-login==0.5.0
flask-opentracing==1.1.0
flask-session==0.4.0
flask-socketio==5.1.1
flask==2.0.1; python_version >= '3.6'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
greenlet==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.2; python_version >= '3'
itsdangerous==2.0.1; python_version >= '3.6'
jaeger-client==4.6.0
jinja2==3.0.1; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
msgpack==1.0.2
opentracing-instrumentation==3.3.1
opentracing==2.4.0
prometheus-client==0.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
prometheus-flask-exporter==0.18.2
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt[crypto]==2.1.0
pylru==1.2.0
python-dotenv==0.19.0
python-engineio==4.2.1; python_version >= '3.6'
python-socketio==5.4.0; python_version >= '3.6'
redis-pubsub-dict==0.0.2
redis-py-cluster==2.1.3
redis==3.5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
requests==2.26.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
threadloop==1.0.2
thrift==0.13.0
tornado==5.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.1; python_version >= '3.6'
wrapt==1.12.1
